# docs/swagger.yaml
basePath: https://api.web95.tech/api/v1
definitions:
  helpers.ErrorData:
    properties:
      code:
        type: string
      fields:
        additionalProperties:
          type: string
        type: object
      message:
        type: string
    type: object
...
  models.CreateReviewRequest:
    properties:
      comment:
        example: Отличный тур!
        type: string
      rating:
        example: 5
        type: integer
      trip_id:
        example: 1
        type: integer
      user_name:
        example: Иван
        type: string
    type: object
...
  models.Order:
    properties:
      created_at:
        type: string
      date:
        type: string
      id:
        type: integer
      is_read:
        type: boolean
      name:
        description: пользовательские поля без префиксов
        type: string
      phone:
        type: string
      price:
        type: string
      status:
        type: string
      trip_id:
        example: 123
        type: integer
      username:
        type: string
    type: object
  models.PaginatedTripReviews:
    properties:
      items:
        items:
          $ref: '#/definitions/models.TripReview'
        type: array
      total:
        type: integer
    type: object
...
  models.TripReview:
    properties:
      comment:
        type: string
      created_at:
        type: string
      id:
        type: integer
      rating:
        type: integer
      trip_id:
        type: integer
      user_name:
        type: string
    type: object
...
  services.OrdersWithTotal:
    properties:
      orders:
        items:
          $ref: '#/definitions/models.Order'
        type: array
      total:
        type: integer
    type: object
info:
  contact: {}
  description: API for Travel project
  title: Travel API
  version: "1.0"
paths:
...
  /admin/orders:
    get:
      description: Получить список заказов с пагинацией и фильтрацией
      parameters:
      - description: Количество записей (по умолчанию 20)
        in: query
        name: limit
        type: integer
      - description: Смещение (по умолчанию 0)
        in: query
        name: offset
        type: integer
      - description: Фильтр по статусу (new/in_progress/done/canceled)
        in: query
        name: status
        type: string
      - description: Фильтр по телефону
        in: query
        name: phone
        type: string
      - description: Фильтр по прочитанности
        in: query
        name: is_read
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.OrdersWithTotal'
        "500":
          description: Не удалось получить список заказов
          schema:
            $ref: '#/definitions/helpers.ErrorData'
      security:
      - Bearer: []
      summary: Get orders list (admin)
      tags:
      - Admin — Orders
  /admin/orders/{id}:
    delete:
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.ErrorData'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.ErrorData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.ErrorData'
      security:
      - Bearer: []
      summary: Delete order
      tags:
      - Admin — Orders
  /admin/orders/{id}/read:
    post:
      description: Пометить заказ как прочитанный
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Некорректный ID
          schema:
            $ref: '#/definitions/helpers.ErrorData'
        "500":
          description: Не удалось обновить заказ
          schema:
            $ref: '#/definitions/helpers.ErrorData'
      security:
      - Bearer: []
      summary: Mark order as read
      tags:
      - Admin — Orders
  /admin/orders/{id}/status:
    post:
      description: Обновить статус заказа (admin)
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      - description: Новый статус (new/in_progress/done/canceled)
        in: query
        name: status
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Некорректные данные
          schema:
            $ref: '#/definitions/helpers.ErrorData'
        "404":
          description: Заказ не найден
          schema:
            $ref: '#/definitions/helpers.ErrorData'
        "500":
          description: Не удалось обновить статус
          schema:
            $ref: '#/definitions/helpers.ErrorData'
      security:
      - Bearer: []
      summary: Update order status
      tags:
      - Admin — Orders
...
  /trips/{trip_id}/reviews:
    get:
      description: Получить список отзывов по туру
      parameters:
      - description: Trip ID
        in: path
        name: trip_id
        required: true
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaginatedTripReviews'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.ErrorData'
      summary: List reviews for trip
      tags:
      - Public — Reviews
    post:
      consumes:
      - application/json
      description: Добавить отзыв к туру
      parameters:
      - description: Review
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateReviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TripReview'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.ErrorData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.ErrorData'
      summary: Leave review
      tags:
      - Public — Reviews
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"